Handle 2-tuples
This is a first cut at handling tuples.  For now, only 2-tuples are supported, but I plan to support up to 5-tuples as that is what pyml supports.
Handle 2-tuples

This is a first cut at handling tuples.  For now, only 2-tuples are supported, but I plan to support up to 5-tuples as that is what pyml supports.